Keep updating the PoS architecture.
